Formatter
- added DOUBLE and VECTOR_RESIDUALS formatter
- reorganisiere package structure
