+++ JAG3D/src/org/applied_geodesy/ui/textfield/UncertaintyTextField.java
@@ -1,0 +1,40 @@
+/***********************************************************************
+* Copyright by Michael Loesler, https://software.applied-geodesy.org   *
+*                                                                      *
+* This program is free software; you can redistribute it and/or modify *
+* it under the terms of the GNU General Public License as published by *
+* the Free Software Foundation; either version 3 of the License, or    *
+* at your option any later version.                                    *
+*                                                                      *
+* This program is distributed in the hope that it will be useful,      *
+* but WITHOUT ANY WARRANTY; without even the implied warranty of       *
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        *
+* GNU General Public License for more details.                         *
+*                                                                      *
+* You should have received a copy of the GNU General Public License    *
+* along with this program; if not, see <http://www.gnu.org/licenses/>  *
+* or write to the                                                      *
+* Free Software Foundation, Inc.,                                      *
+* 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.            *
+*                                                                      *
+***********************************************************************/
+
+package org.applied_geodesy.ui.textfield;
+
+import org.applied_geodesy.util.CellValueType;
+
+public class UncertaintyTextField extends DoubleTextField {
+	
+	public UncertaintyTextField(Double value, CellValueType type) {
+		this(value, type, false);
+	}
+
+	public UncertaintyTextField(Double value, CellValueType type, boolean displayUnit) {
+		this(value, type, displayUnit, ValueSupport.EXCLUDING_INCLUDING_INTERVAL); // GREATER_THAN_OR_EQUAL_TO_ZERO);
+	}
+	
+	public UncertaintyTextField(Double value, CellValueType type, boolean displayUnit, ValueSupport valueSupport) {
+		super(value, type, displayUnit, valueSupport, 0.0, Double.POSITIVE_INFINITY);
+	}
+}
+
